backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9

# CORS middleware
fastapi-cors==0.0.6

# Excel file handling (for migration and exports)
openpyxl==3.1.2
xlsxwriter==3.1.9
numpy==1.26.0
pandas==2.1.3

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# HTTP client for testing
httpx==0.25.2

# Logging and monitoring
loguru==0.7.2
structlog==23.2.0

# Environment variable management
python-dotenv==1.0.0

# Data validation and serialization
typing-extensions==4.8.0

# Redis for caching and rate limiting
redis==5.0.1

# Email with OAuth support
msal==1.25.0  # For Microsoft OAuth

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0